requirements.txt
fastapi
uvicorn[standard]
pytest 
pytest-asyncio
httpx
python-dotenv